+++ Example/LocalImagesExample/PresentedViewController.h
@@ -1,0 +1,13 @@
+//
+//  PresentedViewController.h
+//  PhotoBrowser
+//
+//  Created by Moch Xiao on 5/17/16.
+//  Copyright © 2016 Moch. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface PresentedViewController : UIViewController <UIViewControllerTransitioningDelegate>
+
+@end