+++ CHANGELOG.md
@@ -9,4 +9,4 @@ PhotoBrowser is a light weight photo browser, like the wechat, weibo image viewe
 -   修复 issue #3 
 -   废弃 `setInitializePageIndex:` ，改为 `pb_startPage`。
 -   废弃 `PBViewControllerDataSource` 中的 `viewController:presentImageView:forPageAtIndex:`方法，改为 `viewController:presentImageView:forPageAtIndex:progressHandler`。
--   移除`PBViewControllerDelegate` 中的 `viewController:didSingleTapedPageAtIndex:presentedImage` 方法。
+-   去除 `PBViewControllerDelegate` 中的 `viewController:didSingleTapedPageAtIndex:presentedImage` 退出事件。交由实现者来决定。（例如 Twitter 单击并不退出，而是切换显示工具栏。）